QPID-122
Pulled _readable (_readableMessage) boolean up to AbstractJMSMessage.java and created (_readableProperties).
Checks are now carried out to ensure that the message is readable/writable for the operation so that the correct exception can be thrown. As per the spec.
clearBody() method had been implemented in AbstractJMSMessage.java to reset the _readableMessage as a result implementations must now implement clearBodyImpl() which is called before the _readableMessage is reset.

Test updated to check that properties and body are correctly read/write set.

git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid/java@480119 13f79535-47bb-0310-9956-ffa450edef68