This commit is related to QPID-1161. Please refer to the JIRA for complete details.
In Summary this contains a simple test kit comprising of perf and soak tests.
The focus is on producing a packaged set of tests that can be easily deployed on target environment.

For Quick perf report for a particular release, please run perf_report.sh which will show results for 8 common use cases in a tabular format.



git-svn-id: https://svn.apache.org/repos/asf/incubator/qpid/trunk/qpid/java@672810 13f79535-47bb-0310-9956-ffa450edef68