QPID-2092  Move test configuration to systest package and update to provide an more reliable way of adjusting the configuration form the test. In doing this it made sense to udpate the configurations such that they all inherit from the cannonical config.xml from the broker package. Each test now gets its own configuration file that is used to override the defaut values in config.xml.
So the config is now a hierarchy of:
- Test Specific
- Systest adjustments
- stock broker config.xml

I have disabled the DerbyMessageStoreLoggingTest as after making this change it is more reliably throwing a Derby exception on broker startup.

The other changes to the tests were to ensure they correctly save the Test Specific config file before loading. As for some reason, Commons Configuration is not abiding by the optional flag.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@813461 13f79535-47bb-0310-9956-ffa450edef68