QPID-2625 : Moved Logging generation to moudule.xml to allow plugins to utilise the same functionality. To enable generation for your plugin just add : <target name="precompile" depends="gen_logging"/> to your build.xml
Logging is now defined in a X_logmessage.properties file. Where X is used to make the XMessages.java class.

Also updated all existing usages to remove the 3 digit prefix that wasn't adding any info.
Updated ConfigStore and Transaction Log to use named properties rather than numeric values.

If we are going to continue with <3 alpha>-<4 numeric> ids for messages then we'll need to have some registry to prevent
clases. Perhaps it is simpler to relax this and require a plugin creator to provide a unique identifier for their messages.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@954432 13f79535-47bb-0310-9956-ffa450edef68