QPID-3865: Remove unreliable CancelTest
This test case is seen to fail occasionally on a slower CI box. The failure is a test issue; it uses a non-transactional producer to send the single test message, so the message may not have made it to the broker by the time the QueueBrowser is created. When the test fails, the browser sees no messages.

The test won't be fixed as the test is duplicated by the better QueueBrowser*Test which uses a transactional publisher.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@1292207 13f79535-47bb-0310-9956-ffa450edef68