QPID-3599: Python session tests fail against the Java Broker due to unexpected redelivered status
Changed Java Broker so once a message is acknowledged, if the session is transactional and the
transaction is rolled-back, the queue entry is marked as redelivered.  The rationale is that if the
client has acknowledged, then it is certain they have seen it/them.

The python tests testCommitAck and testRollback highlighted this defect as they acknowledge the session
before rollback (something that is impossible to control from the JMS API).

Python tests reenabled against the Java Broker:
qpid.tests.messaging.endpoints.SessionTests.testCommitAck
qpid.tests.messaging.endpoints.SessionTests.testRelease
qpid.tests.messaging.endpoints.SessionTests.testRollback

Work of Robbie Gemmell <robbie@apache.org> and myself.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@1295474 13f79535-47bb-0310-9956-ffa450edef68