QPID-4503: Producer transaction timeout detection feature may produce suprious open/idle alerts and close client connections/sessions without good cause
Race conditon existed between the initial check that determined transaction timeout detection was required and the subsequent re-observeration
required to perform the open and idle calculation.  In the unlucky timing, the state of the transaction changed between these two points.  (To produce
the time-since-epoch type the transaction needs to be committed between these two points).

Changed checkTransactionStatus so that transactionStartTime and transactionUpdateTime are observed once only.

There still exists the possibility that the transaction status change change between the reading of transactionStartTime and transactionUpdateTime
times, but I do not see how this could produce a suprious report (or close).  The alternative (to take a locks to ensure consistent observations
are made) seems unjustifiably costly.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk/qpid/java@1421884 13f79535-47bb-0310-9956-ffa450edef68